Update to delete outputs and put into one file
Set it up so that  there are four vectors. Two hold the word and count of the 'master' and two hold the word and count of the other files to work through. Two for loops to compare if words are ever the same and add the counts if they are. Erase any matches and then at the end add all remaining words and counts to the master. Seems to go pretty quick and I compared the numbers to our phase 1 and everything seems to be fine.